nsxwidget: add block content feature
The main motivation of this patch is to add a way for the user to
disable loading any external resources so that the user can prevent
privacy tracking type html pages. For the most part, this use-case is
from viewing html emails with xwidgets. Theoretically, this can provide
more advanced filter rules since the user can pass a string of their own
custom rules (e.g. only block css resources).

For more information on content rules, see [1].

This patch only implements the nextstep version of it.

[1]: https://developer.apple.com/documentation/safariservices/creating_a_content_blocker